+++ CHANGELOG.md
@@ -75,6 +75,7 @@
 55. Fixed Issue with Reflection based Assembly Resolving.
 56. Moved Plugins and Mods folders back to the Game Installation folder.
 57. Fixed ConsoleOnTop Focusing Issues.
+58. Added More HostNames to Analytics Blocker.
 
 ---